+++ source/isaaclab_rl/test/test_skrl_wrapper.py
@@ -18,9 +18,8 @@ import gymnasium as gym
 import pytest
 import torch
 
-import carb
-
 import isaaclab.sim as sim_utils
+from isaaclab.app.settings_manager import get_settings_manager
 from isaaclab.envs import DirectMARLEnv, multi_agent_to_single_agent
 
 from isaaclab_rl.skrl import SkrlVecEnvWrapper
@@ -44,8 +43,7 @@ def registered_tasks():
 
     # this flag is necessary to prevent a bug where the simulation gets stuck randomly when running the
     # test on many environments.
-    carb_settings_iface = carb.settings.get_settings()
-    carb_settings_iface.set_bool("/physics/cooking/ujitsoCollisionCooking", False)
+    get_settings_manager().set_bool("/physics/cooking/ujitsoCollisionCooking", False)
 
     # print all existing task names
     print(">>> All registered environments:", registered_tasks)
@@ -63,7 +61,7 @@ def test_random_actions(registered_tasks):
         # create a new stage
         sim_utils.create_new_stage()
         # reset the rtx sensors carb setting to False
-        carb.settings.get_settings().set_bool("/isaaclab/render/rtx_sensors", False)
+        get_settings_manager().set_bool("/isaaclab/render/rtx_sensors", False)
         try:
             # parse configuration
             env_cfg = parse_env_cfg(task_name, device=device, num_envs=num_envs)